requirements.txt
# 核心依赖
openai>=1.0.0
jieba>=0.42.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# CLI 界面
rich>=13.0.0
typer>=0.9.0

# 数据可视化
networkx>=3.0
matplotlib>=3.7.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 文档生成
python-docx>=0.8.11
marko>=2.0.0

# 音频处理（可选）
whisper-openai>=20231117

# HTTP 请求（用于 Ollama 本地模型）
requests>=2.31.0

# 测试依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
httpx>=0.25.0  # 用于测试 HTTP 请求

# 开发工具（可选）
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
